Use batch_size_per_device
